+++ ui/StatusQ/src/keychain.cpp
@@ -5,10 +5,6 @@
 Keychain::Keychain(QObject *parent) : QObject(parent)
 {}
 
-#ifndef Q_OS_MACOS
-Keychain::~Keychain() = default;
-#endif
-
 QString Keychain::service() const
 {
     return m_service;
@@ -50,32 +46,3 @@ void Keychain::setLoading(bool loading)
     m_loading = loading;
     emit loadingChanged();
 }
-
-#ifndef Q_OS_MACOS
-void Keychain::requestSaveCredential(const QString& account, const QString& password)
-{
-    Q_UNUSED(account);
-    Q_UNUSED(password);
-    qWarning() << "Keychain::requestSaveCredential is intended to be called only on MacOS.";
-    emit saveCredentialRequestCompleted(Keychain::StatusNotSupported);
-}
-
-void Keychain::requestDeleteCredential(const QString& account)
-{
-    Q_UNUSED(account);
-    qWarning() << "Keychain::requestDeleteCredential is intended to be called only on MacOS.";
-    emit deleteCredentialRequestCompleted(Keychain::StatusNotSupported);
-}
-
-void Keychain::requestGetCredential(const QString& account)
-{
-    Q_UNUSED(account);
-    qWarning() << "Keychain::requestGetCredential is intended to be called only on MacOS.";
-    emit getCredentialRequestCompleted(Keychain::StatusNotSupported, "");
-}
-
-void Keychain::cancelActiveRequest()
-{
-    qWarning() << "Keychain::cancelActiveRequest is intended to be called only on MacOS.";
-}
-#endif
