+++ ui/StatusQ/src/keychain_other.cpp
@@ -1,0 +1,30 @@
+#include "StatusQ/keychain.h"
+
+Keychain::~Keychain() = default;
+
+void Keychain::requestSaveCredential(const QString &account, const QString &password)
+{
+    Q_UNUSED(account);
+    Q_UNUSED(password);
+    qWarning() << "Keychain::requestSaveCredential is intended to be called only on MacOS.";
+    emit saveCredentialRequestCompleted(Keychain::StatusNotSupported);
+}
+
+void Keychain::requestDeleteCredential(const QString &account)
+{
+    Q_UNUSED(account);
+    qWarning() << "Keychain::requestDeleteCredential is intended to be called only on MacOS.";
+    emit deleteCredentialRequestCompleted(Keychain::StatusNotSupported);
+}
+
+void Keychain::requestGetCredential(const QString &account)
+{
+    Q_UNUSED(account);
+    qWarning() << "Keychain::requestGetCredential is intended to be called only on MacOS.";
+    emit getCredentialRequestCompleted(Keychain::StatusNotSupported, "");
+}
+
+void Keychain::cancelActiveRequest()
+{
+    qWarning() << "Keychain::cancelActiveRequest is intended to be called only on MacOS.";
+}
